Add AWS S3 storage compatibility for encryption/decryption
Agent-Logs-Url: https://github.com/jemysh007/carrierwave_encrypter_decrypter/sessions/b3857cd1-5b5a-4553-98ef-1107635cd193

+++ lib/carrierwave/encrypter_decrypter/storage_helper.rb
@@ -2,33 +2,66 @@ module Carrierwave
   module EncrypterDecrypter
     module StorageHelper
       class << self
+        def remote_storage?(uploader)
+          fog_storage?(uploader) || aws_storage?(uploader)
+        end
+
         def fog_storage?(uploader)
           defined?(CarrierWave::Storage::Fog) &&
-            uploader.class.storage == CarrierWave::Storage::Fog
+            storage_class(uploader) == CarrierWave::Storage::Fog
         rescue
           false
         end
 
-        def fog_read(uploader, key)
-          file = fog_directory(uploader).files.get(key)
-          raise "File not found in fog storage: #{key}" unless file
-          file.body
+        def aws_storage?(uploader)
+          defined?(CarrierWave::Storage::AWS) &&
+            storage_class(uploader) == CarrierWave::Storage::AWS
+        rescue
+          false
         end
 
-        def fog_write(uploader, key, data)
-          fog_directory(uploader).files.create(
-            key: key,
-            body: data,
-            public: fog_public(uploader)
-          )
+        def read(uploader, key)
+          if fog_storage?(uploader)
+            file = fog_directory(uploader).files.get(key)
+            raise "File not found in fog storage: #{key}" unless file
+            file.body
+          elsif aws_storage?(uploader)
+            aws_object(uploader, key).get.body.read
+          else
+            raise "Unsupported remote storage for read: #{storage_class(uploader)}"
+          end
         end
 
-        def fog_delete(uploader, key)
-          file = fog_directory(uploader).files.get(key)
-          file.destroy if file
+        def write(uploader, key, data)
+          if fog_storage?(uploader)
+            fog_directory(uploader).files.create(
+              key: key,
+              body: data,
+              public: fog_public(uploader)
+            )
+          elsif aws_storage?(uploader)
+            aws_object(uploader, key).put(body: data)
+          else
+            raise "Unsupported remote storage for write: #{storage_class(uploader)}"
+          end
+        end
+
+        def delete(uploader, key)
+          if fog_storage?(uploader)
+            file = fog_directory(uploader).files.get(key)
+            file.destroy if file
+          elsif aws_storage?(uploader)
+            aws_object(uploader, key).delete
+          else
+            raise "Unsupported remote storage for delete: #{storage_class(uploader)}"
+          end
         end
 
         private
+
+        def storage_class(uploader)
+          uploader.class.storage
+        end
 
         def fog_connection(uploader)
           ::Fog::Storage.new(uploader.class.fog_credentials)
@@ -42,6 +75,19 @@ module Carrierwave
           uploader.class.fog_public
         rescue
           false
+        end
+
+        def aws_object(uploader, key)
+          raise "AWS SDK not loaded" unless defined?(::Aws::S3::Object)
+          file = uploader.file
+          raise "Uploader file is unavailable for AWS storage" unless file
+
+          object = file.file
+          unless object.is_a?(::Aws::S3::Object)
+            raise "Unsupported AWS file object: #{object.class}"
+          end
+
+          object.bucket.object(key)
         end
       end
     end
